feat(automation): Add GitHub Action to handle creating a new release
This commit replaces the "placeholder" create-release.yaml file to include functionality to actually create a opendatahub-io/kubeflow release following the original manual process.

The GHA can be invoked manually as a workflow from GitHub as well as invoked as part of the kubeflow-release action.

`kubeflow-release.yaml` has been modified to pass in the target branch name for cutting the release.  The release/tag name is **NOT** provided as the `create-release.sh` script will auto-compute what the proper tag name should be.
- Please note that the tag name **can** be provided if necessary in the future... at which point the tag name would be honored as is (i.e. no "auto-incrementing" logic is performed if the tag name provided as input)

`create-release.sh` will also handle auto-updating the release tag for future scenarios where we change the target branch (ex: `v1.9-branch` to `v2.0-branch`).

Please see the (extensive) comments in `.github/scripts/create-release.sh` for deeper explanation on implementation.
- Please note, in what may be a "controversial" decision, I have extracted the script into its own file as opposed to embedding it within the `yaml` of the action.  I find it more pleasant to read/develop when the script is in its own file - and also allows code editors to be smarter in offering hints on the source.

Related-to: https://issues.redhat.com/browse/RHOAIENG-15391